Moved to shellExecutionHelper
