+++ extensions/sql-database-projects/src/models/deploy/deployService.ts
@@ -14,6 +14,7 @@ import * as vscode from 'vscode';
 import * as os from 'os';
 import { ConnectionResult } from 'azdata';
 import * as templates from '../../templates/templates';
+import { ShellExecutionHelper } from '../../tools/shellExecutionHelper';
 
 interface DockerImageSpec {
 	label: string;
@@ -92,7 +93,7 @@ export class DeployService {
 
 	private async verifyDocker(): Promise<void> {
 		try {
-			await utils.executeCommand(`docker version --format {{.Server.APIVersion}}`, this._outputChannel);
+			await ShellExecutionHelper.executeCommand(`docker version --format {{.Server.APIVersion}}`, this._outputChannel);
 			// TODO verify min version
 		} catch (error) {
 			throw Error(constants.dockerNotRunningError(utils.getErrorMessage(error)));
@@ -170,7 +171,7 @@ export class DeployService {
 			// Waiting a bit to make sure docker container doesn't crash
 			//
 			const runningDockerId = await utils.retry('Validating the docker container', async () => {
-				return await utils.executeCommand(`docker ps -q -a --filter label=${imageSpec.label} -q`, this._outputChannel);
+				return await ShellExecutionHelper.executeCommand(`docker ps -q -a --filter label=${imageSpec.label} -q`, this._outputChannel);
 			}, (dockerId) => {
 				return Promise.resolve({ validated: dockerId !== undefined, errorMessage: constants.dockerContainerNotRunningErrorMessage });
 			}, (dockerId) => {
@@ -186,7 +187,7 @@ export class DeployService {
 				this.logToOutput(constants.dockerContainerFailedToRunErrorMessage);
 				if (createdDockerId) {
 					// Get the docker logs if docker was created but crashed
-					await utils.executeCommand(constants.dockerLogMessage(createdDockerId), this._outputChannel);
+					await ShellExecutionHelper.executeCommand(constants.dockerLogMessage(createdDockerId), this._outputChannel);
 				}
 			}
 
@@ -201,13 +202,13 @@ export class DeployService {
 
 		// Running commands to build the docker image
 		this.logToOutput('Building docker image ...');
-		await utils.executeCommand(`docker pull ${profile.dockerBaseImage}`, this._outputChannel);
-		await utils.executeCommand(`docker build -f ${dockerFilePath} -t ${dockerImageSpec.tag} ${root}`, this._outputChannel);
-		await utils.executeCommand(`docker images --filter label=${dockerImageSpec.label}`, this._outputChannel);
+		await ShellExecutionHelper.executeCommand(`docker pull ${profile.dockerBaseImage}`, this._outputChannel);
+		await ShellExecutionHelper.executeCommand(`docker build -f ${dockerFilePath} -t ${dockerImageSpec.tag} ${root}`, this._outputChannel);
+		await ShellExecutionHelper.executeCommand(`docker images --filter label=${dockerImageSpec.label}`, this._outputChannel);
 
 		this.logToOutput('Running docker container ...');
-		await utils.executeCommand(`docker run -p ${profile.port}:1433 -e "MSSQL_SA_PASSWORD=${profile.password}" -d --name ${dockerImageSpec.containerName} ${dockerImageSpec.tag}`, this._outputChannel, sensitiveData);
-		return await utils.executeCommand(`docker ps -q -a --filter label=${dockerImageSpec.label} -q`, this._outputChannel);
+		await ShellExecutionHelper.executeCommand(`docker run -p ${profile.port}:1433 -e "MSSQL_SA_PASSWORD=${profile.password}" -d --name ${dockerImageSpec.containerName} ${dockerImageSpec.tag}`, this._outputChannel, sensitiveData);
+		return await ShellExecutionHelper.executeCommand(`docker ps -q -a --filter label=${dockerImageSpec.label} -q`, this._outputChannel);
 	}
 
 	private async getConnectionString(connectionUri: string): Promise<string | undefined> {
@@ -379,7 +380,7 @@ export class DeployService {
 		//
 		await this.createFile(startFilePath, 'echo starting the container!');
 		if (os.platform() !== 'win32') {
-			await utils.executeCommand(`chmod +x '${startFilePath}'`, this._outputChannel);
+			await ShellExecutionHelper.executeCommand(`chmod +x '${startFilePath}'`, this._outputChannel);
 		}
 
 		// Create the Dockerfile
@@ -402,7 +403,7 @@ RUN ["/bin/bash", "/opt/commands/start.sh"]
 	}
 
 	public async getCurrentDockerContainer(imageLabel: string): Promise<string[]> {
-		const currentIds = await utils.executeCommand(`docker ps -q -a --filter label=${imageLabel}`, this._outputChannel);
+		const currentIds = await ShellExecutionHelper.executeCommand(`docker ps -q -a --filter label=${imageLabel}`, this._outputChannel);
 		return currentIds ? currentIds.split(/\r?\n/) : [];
 	}
 
@@ -412,7 +413,7 @@ RUN ["/bin/bash", "/opt/commands/start.sh"]
 			if (id) {
 				for (let commandId = 0; commandId < commandsToClean.length; commandId++) {
 					const command = commandsToClean[commandId];
-					await utils.executeCommand(`${command} ${id}`, this._outputChannel);
+					await ShellExecutionHelper.executeCommand(`${command} ${id}`, this._outputChannel);
 				}
 			}
 		}
